% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_parens}
\alias{add_parens}
\title{Add parentheses around standard error estimates}
\usage{
add_parens(x, digits = 3)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of digits to retain}
}
\value{
A character vector with enclosing parentheses
}
\description{
Add parentheses around standard error estimates
}
\examples{

std.error <- c(0.12, 0.001, 1.2)
add_parens(std.error)

}
