% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch18_helpers.R
\name{best_predictor}
\alias{best_predictor}
\title{Best predictor function from causal_forest}
\usage{
best_predictor(data, covariate_names, cuts = 20)
}
\arguments{
\item{data}{A data.frame of covariates}

\item{covariates}{A character vector of covariates to assess}
}
\value{
a data.frame of the best predictors
}
\description{
Best predictor function from causal_forest
}
