% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bonilla_tillery}
\alias{bonilla_tillery}
\title{Replication data for Bonilla and Tillery (2020), American Political Science Review (obtained from Dataverse 10.7910/DVN/IUZDQI)}
\format{
A data.frame
}
\usage{
bonilla_tillery
}
\description{
Replication data for Bonilla and Tillery (2020), American Political Science Review (obtained from Dataverse 10.7910/DVN/IUZDQI)
}
\keyword{datasets}
