% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch18_helpers.R
\name{causal_forest_handler}
\alias{causal_forest_handler}
\title{Tidy causal forest results}
\usage{
causal_forest_handler(data, covariates, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{Options to causal_forest}
}
\description{
Tidy causal forest results
}
