% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch18_helpers.R
\name{causal_forest_helper}
\alias{causal_forest_helper}
\title{Tidy helper function for causal_forest function}
\usage{
causal_forest_helper(data, covariates, ...)
}
\arguments{
\item{data}{A data.frame}

\item{covariates}{A character vector of covariates to assess}

\item{...}{Options to causal_forest}
}
\description{
Runs estimates estimation function from interference package and returns tidy data frame output
}
\details{
https://draft.declaredesign.org/complex-designs.html#discovery-using-causal-forests

See ?causal_forest for further details
}
