% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11_helpers.R
\name{cef_inquiry}
\alias{cef_inquiry}
\title{Conditional expectation function inquiry}
\usage{
cef_inquiry(data)
}
\arguments{
\item{data}{A data.frame (ignored)}
}
\value{
A tibble (data.frame) with the X value and estimand value for the dip conditional expectation function
}
\description{
Constructs a conditional expectation function over the range 0 to 3 for illustration nonstandard inquiries
}
\details{
See Ch. 11 https://book.declaredesign.org/redesign.html
}
