% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{conjoint_assignment}
\alias{conjoint_assignment}
\title{Conjoint experiment assignment handler: conducts complete random assignment of all attribute levels}
\usage{
conjoint_assignment(data, levels_list)
}
\arguments{
\item{levels_list}{}
}
\value{

}
\description{
Conjoint experiment assignment handler: conducts complete random assignment of all attribute levels
}
