% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dd_palette}
\alias{dd_palette}
\title{Access color palette used in the book "Research Design: Declare, Diagnose, Redesign" (Blair, Coppock, Humphreys)}
\usage{
dd_palette(name, n)
}
\arguments{
\item{name}{Color palette name (character)}

\item{n}{Number of colors}
}
\value{
character vector of colors
}
\description{
Based on Karthik Ram's wesanderson package (https://github.com/karthik/wesanderson)
}
\details{
Available color palettes:

color_palette = c("#72B4F3", "#F38672", "#C6227F")

grey_palette = c("#72B4F3", "#F38672", "#C6227F", gray(0.8))

dd_dark_blue = "#3564ED"

dd_light_blue = "#72B4F3"

dd_orange = "#F38672"

dd_purple = "#7E43B6"

dd_gray = gray(0.2)

dd_pink = "#C6227F"

dd_light_gray = gray(0.8)

dd_dark_blue_alpha = "#3564EDA0"

dd_light_blue_alpha = "#72B4F3A0"
}
