% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15_helpers.R
\name{did_multiplegt_tidy}
\alias{did_multiplegt_tidy}
\title{Tidy estimates from did_multiplegt}
\usage{
did_multiplegt_tidy(data, ...)
}
\arguments{
\item{data}{a data.frame}

\item{...}{options passed to did_multiplegt}
}
\description{
See https://book.declaredesign.org/observational-causal.html#difference-in-differences
}
