% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11_helpers.R
\name{dip}
\alias{dip}
\title{Dip CEF function}
\usage{
dip(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
a numeric vector
}
\description{
"Dip" shaped functional form for illustration of nonstandard inquiries
}
\details{
See Ch. 11 https://book.declaredesign.org/redesign.html
}
