% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch17_helpers.R
\name{estimator_AS}
\alias{estimator_AS}
\title{Helper function for the Aronow and Samii estimator}
\usage{
estimator_AS(data, permutatation_matrix, adj_matrix)
}
\arguments{
\item{data}{a data.frame}

\item{permutatation_matrix}{a permuatation matrix of random assignments}

\item{adj_matrix}{an adjacency matrix}
}
\description{
See https://book.declaredesign.org/experimental-causal.html#experiments-over-networks
}
