% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fairfax}
\alias{fairfax}
\title{Shapefile of Fairfax County, Virginia, voting precincts}
\format{
An sf object with 236 rows and 10 variables:
}
\usage{
fairfax
}
\description{
An sf object containing the boundaries of voting precincts for Fairfax County, Virginia as well as precinct ID, name, district, polling place name, address, city, zip code, area, length, and geometry (polygons)
}
\keyword{datasets}
