% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{hex_add_alpha}
\alias{hex_add_alpha}
\title{Add alpha transparency to a color defined in hexadecimal}
\usage{
hex_add_alpha(col, alpha)
}
\arguments{
\item{col}{Original color code in hex}

\item{alpha}{Level of alpha transparency to add}
}
\value{
color codes with alpha added
}
\description{
Add alpha transparency to a color defined in hexadecimal
}
