% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_voter_file}
\alias{la_voter_file}
\title{Voter file sample for Los Angeles County}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
\item{party}{political party registration}
\item{age}{age of voter in years}
\item{census_tract}{US Census tract number}
\item{voted_2012}{voter turnout in 2012 election}
}
}
\source{
California Secretary of State.
}
\usage{
la_voter_file
}
\description{
A dataset containing the party registration, age, census tract number, and voter turnout in 2012 for 1,000 randomly-sampled registered voters in Los Angeles County, California.
}
\keyword{datasets}
