% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-difference-in-differences.R
\name{lag_by_group}
\alias{lag_by_group}
\title{Generate lags in grouped data}
\usage{
lag_by_group(x, groups, n = 1, order_by, default = NA)
}
\arguments{
\item{x}{Vector of values}

\item{groups}{Grouping variable}

\item{n}{Positive integer of length 1, giving the number of positions to lead or lag by}

\item{order_by}{Ordering variable withing group (e.g., time)}

\item{default}{Value used for non-existent rows. Defaults to NA.}
}
\value{
vector of lagged values
}
\description{
See https://book.declaredesign.org/observational-causal.html#difference-in-differences
}
