% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lapop_brazil}
\alias{lapop_brazil}
\title{Data used in student exercises for RDSS from LAPOP survey of Brazil in 2018}
\format{
A data.frame
}
\usage{
lapop_brazil
}
\description{
See https://www.vanderbilt.edu/lapop/core-surveys.php for survey questionnaire
}
\keyword{datasets}
