% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_interval_entry}
\alias{make_interval_entry}
\title{Format confidence intervals for nice printing}
\usage{
make_interval_entry(conf.low, conf.high, digits = 2)
}
\arguments{
\item{conf.low}{a numeric vector of lower bounds}

\item{conf.high}{a numeric vector of upper bounds}

\item{digits}{number of digits to retain}
}
\value{
a character vector of intervals
}
\description{
Format confidence intervals for nice printing
}
\examples{

conf.low <- c(-0.1652, 0.00304, -6.352)
conf.high <- c(0.3052, 0.00696, -1.648)

make_interval_entry(conf.low, conf.high)


}
