% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_se_entry}
\alias{make_se_entry}
\title{Format estimates and standard errors for nice printing}
\usage{
make_se_entry(estimate, std.error, digits = 2)
}
\arguments{
\item{estimate}{a numeric vector of parameter estimates}

\item{std.error}{a numeric vector of standard error estimates}

\item{digits}{number of digits to retain}
}
\value{
a character vector of formatted estimates and standard errors
}
\description{
Format estimates and standard errors for nice printing
}
\examples{

estimate <- c(0.07, 0.005, -4)
std.error <- c(0.12, 0.001, 1.2)

make_se_entry(estimate, std.error)

}
