% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch18_helpers.R
\name{mle2_helper}
\alias{mle2_helper}
\title{Helper function for mle2 function in bbmle package}
\usage{
mle2_helper(data, optimization_method, ...)
}
\arguments{
\item{data}{a data.frame}

\item{optimization_method}{Optimization method to use. See optim.}

\item{...}{Further arguments to pass to mle2

Note that because DeclareDesign uses the \code{optimization_method} argument in \code{declare_estimator}, the type method is renamed from \code{method} used in \code{rma}.)}
}
\value{
an mle2 object
}
\description{
See https://draft.declaredesign.org/experimental-descriptive.html#behavioral-games
}
\details{
See ?mle2 for further details
}
