% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdrobust.R
\name{rdrobust_helper}
\alias{rdrobust_helper}
\title{Tidy helper function for rdrobust function}
\usage{
rdrobust_helper(data, y, x, c, subset = NULL, term = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{y}{is the dependent variable (rdrobust argument)}

\item{x}{is the running variable (a.k.a. score or forcing variable) (rdrobust argument)}

\item{c}{specifies the RD cutoff in x; default is c = 0 (rdrobust argument)}

\item{subset}{An optional bare (unquoted) expression specifying a subset of observations to be used}

\item{term}{Symbols or literal character vector of term that represent quantities of interest, i.e. Z. If FALSE, return the first non-intercept term; if TRUE return all term. To escape non-standard-evaluation use !!.}
}
\description{
Runs rdrobust estimation function and returns tidy data frame output
}
\details{
See https://book.declaredesign.org/observational-causal.html#regression-discontinuity-designs
}
