\name{rdss-package}
\alias{rdss-package}
\alias{rdss}
\docType{package}
\title{
  \packageTitle{rdss}
}
\description{
  \packageDescription{rdss}
}
\author{
\packageAuthor{rdss}

Maintainer: Frédéric Santos, \email{frederic.santos@u-bordeaux.fr}
}
\references{
  Murail, P., Bruzek, J. and Braga, J. (1999) A new approach
  to sexual diagnosis in past populations. Practical adjustments from
  Van Vark's procedure. \emph{International Journal of
  Osteoarchaeology}, \bold{9}, 39--53. doi:
  10.1002/(SICI)1099-1212(199901/02)9:1<39::AID-OA458>3.0.CO;2-V
}
