% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch18_helpers.R
\name{rma_estimator}
\alias{rma_estimator}
\title{Tidy wrapper for rma function in metafor package}
\usage{
rma_estimator(data, yi, sei, method = "REML")
}
\arguments{
\item{data}{a data.frame}

\item{yi}{unquoted variable name of estimates used in meta-analysis}

\item{sei}{unquoted variable name of standard errors used in meta-analysis}

\item{method}{character string to specify whether a fixed- or a random/mixed-effects model should be fitted. A fixed-effects model (with or without moderators) is fitted when using method = "FE". Random/mixed-effects models are fitted by setting method equal to one of the following: "DL", "HE", "SJ", "ML", "REML", "EB", "HS", "HSk", or "GENQ". Default is "REML".}
}
\value{

}
\description{
See https://book.declaredesign.org/complex-designs.html#meta-analysis
}
\details{
See ?rma for further details
}
