% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-meta-analysis.R
\name{rma_mu_tau}
\alias{rma_mu_tau}
\title{Extract mu and tau parameters from rma model fit}
\usage{
rma_mu_tau(fit)
}
\arguments{
\item{fit}{Fit object from the rma function in the metafor package}
}
\value{
a data.frame of estimates
}
\description{
See https://book.declaredesign.org/complex-designs.html#meta-analysis
}
