% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{rr_forced_known}
\alias{rr_forced_known}
\title{Tidy estimates from rrreg}
\usage{
rr_forced_known(formula, data)
}
\arguments{
\item{data}{a data.frame}
}
\description{
See https://draft.declaredesign.org/experimental-descriptive.html#list-experiments
}
