% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{rr_predict_tidy}
\alias{rr_predict_tidy}
\title{Tidy predictions from rrreg}
\usage{
rr_predict_tidy(fit)
}
\arguments{
\item{fit}{a rrreg fit object}
}
\description{
Runs prediction based on rrreg model fit and returns tidy data frame output
}
\details{
See https://draft.declaredesign.org/experimental-descriptive.html#list-experiments
}
