% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dd.R
\name{theme_dd}
\alias{theme_dd}
\title{ggplot Theme used in the book "Research Design: Declare, Diagnose, Redesign" (Blair, Coppock, Humphreys)}
\usage{
theme_dd()
}
\value{
ggplot theme
}
\description{
ggplot Theme used in the book "Research Design: Declare, Diagnose, Redesign" (Blair, Coppock, Humphreys)
}
