% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{tidy.amce}
\alias{tidy.amce}
\title{Tidy estimates from the amce estimator}
\usage{
\method{tidy}{amce}(x, ...)
}
\arguments{
\item{x}{an amce fit object from cjoint::amce}

\item{...}{Extra arguments to pass to tidy}
}
\description{
Runs amce estimation function and returns tidy data frame output
}
\details{
See https://book.declaredesign.org/experimental-descriptive.html#conjoint-experiments
}
\examples{

# data(immigrationconjoint)
# data(immigrationdesign)

# # Run AMCE estimator using all attributes in the design
# results <- amce(Chosen_Immigrant ~  Gender + Education + `Language Skills` +
#                   `Country of Origin` + Job + `Job Experience` + `Job Plans` +
#                   `Reason for Application` + `Prior Entry`, data = immigrationconjoint,
#                 cluster = TRUE, respondent.id = "CaseID", design = immigrationdesign)

# # Print summary
# tidy(results)

}
