% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{tidy.rdrobust}
\alias{tidy.rdrobust}
\title{Tidy helper function for rdrobust function}
\usage{
\method{tidy}{rdrobust}(x, ...)
}
\arguments{
\item{x}{Model fit object from rdrobust}

\item{...}{Other arguments (not used)}
}
\value{
a data.frame of estimates
}
\description{
Runs rdrobust estimation function and returns tidy data frame output
}
\details{
See https://book.declaredesign.org/observational-causal.html#regression-discontinuity-designs
}
