% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9_helpers.R
\name{tidy.stanreg}
\alias{tidy.stanreg}
\title{Tidy results from a stanreg regresion and exponentiate the estimated coefficient}
\usage{
\method{tidy}{stanreg}(fit, conf.int = FALSE, conf.level = 0.95, exponentiate = FALSE, ...)
}
\arguments{
\item{fit}{A stanreg fit from stan_glm}
}
\value{
data.frame of results
}
\description{
Note no standard errors or other summary statistics are provided
}
\details{
See https://book.declaredesign.org/choosing-an-answer-strategy.html#bayesian-formalizations
}
