% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{tidy_amce}
\alias{tidy_amce}
\title{Tidy estimates from the amce estimator}
\usage{
tidy_amce(fit, alpha = 0.05)
}
\arguments{
\item{fit}{an amce fit object from cjoint::amce}

\item{alpha}{the type 1 error rate, set to 0.05 by default. Used for the calucluation of p.values and confidence intervals}
}
\value{

}
\description{
Tidy estimates from the amce estimator
}
