% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9_helpers.R
\name{tidy_exponentiate}
\alias{tidy_exponentiate}
\title{Tidy results from a stanreg regresion and exponentiate the estimated coefficient}
\usage{
tidy_exponentiate(fit)
}
\arguments{
\item{fit}{A stanreg fit from stan_glm}
}
\description{
Note no standard errors or other summary statistics are provided
}
\details{
See https://book.declaredesign.org/choosing-an-answer-strategy.html#bayesian-formalizations
}
