% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch18_helpers.R
\name{best_predictor}
\alias{best_predictor}
\title{Best predictor function from causal_forest}
\usage{
best_predictor(data, covariates)
}
\arguments{
\item{data}{A data.frame}

\item{covariates}{A character vector of covariates to assess}
}
\description{
Best predictor function from causal_forest
}
