% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-causal-forests.R
\name{best_predictor}
\alias{best_predictor}
\title{Best predictor function from causal_forest}
\usage{
best_predictor(data, covariate_names, cuts = 20)
}
\arguments{
\item{data}{A data.frame of covariates}

\item{covariate_names}{A character vector of covariates to assess}

\item{cuts}{Either a numeric vector of two or more unique cut points or a single number (greater than or equal to 2) giving the number of intervals into which each covariate is to be cut.}
}
\value{
a data.frame of the best predictors
}
\description{
Best predictor function from causal_forest
}
