% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11_helpers.R
\name{cef_estimator}
\alias{cef_estimator}
\title{Conditional expectation function estimator}
\usage{
cef_estimator(data)
}
\arguments{
\item{data}{A data.frame}
}
\description{
See Ch. 11 https://book.declaredesign.org/redesign.html
}
