% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11_helpers.R
\name{cef_inquiry}
\alias{cef_inquiry}
\title{Conditional expectation function inquiry}
\usage{
cef_inquiry(data)
}
\arguments{
\item{data}{A data.frame}
}
\description{
See Ch. 11 https://book.declaredesign.org/redesign.html
}
