% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conjoint-experiments.R
\name{conjoint_assignment}
\alias{conjoint_assignment}
\title{Conjoint experiment assignment handler: conducts complete random assignment of all attribute levels}
\usage{
conjoint_assignment(data, levels_list)
}
\arguments{
\item{data}{A data.frame}

\item{levels_list}{List of conjoint levels to assign}
}
\value{
a data.frame with random assignment added
}
\description{
See https://book.declaredesign.org/experimental-descriptive.html#conjoint-experiments
}
