% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{conjoint_inquiries}
\alias{conjoint_inquiries}
\title{Conjoint experiment Inquiries handler}
\usage{
conjoint_inquiries(data, levels_list)
}
\arguments{
\item{levels_list}{}
}
\value{

}
\description{
Conjoint experiment Inquiries handler
}
