% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conjoint-experiments.R
\name{conjoint_inquiries}
\alias{conjoint_inquiries}
\title{Conjoint experiment inquiries handler}
\usage{
conjoint_inquiries(data, levels_list, utility_fn)
}
\arguments{
\item{data}{A data.frame}

\item{levels_list}{List of conjoint levels}

\item{utility_fn}{a function that takes data and returns an additional column called U, which represents the utility of the choice}
}
\value{
a data.frame of estimand values
}
\description{
See https://book.declaredesign.org/experimental-descriptive.html#conjoint-experiments
}
