% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{conjoint_measurement}
\alias{conjoint_measurement}
\title{Conjoint experiment assignment handler: conducts complete random assignment of all attribute levels}
\usage{
conjoint_measurement(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Conjoint experiment assignment handler: conducts complete random assignment of all attribute levels
}
