% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch16_helpers.R
\name{conjoint_measurement}
\alias{conjoint_measurement}
\title{Conjoint experiment assignment handler: conducts complete random assignment of all attribute levels}
\usage{
conjoint_measurement(data, utility_fn)
}
\arguments{
\item{data}{A data.frame}
}
\value{
a data.frame with XXX
}
\description{
See https://book.declaredesign.org/experimental-descriptive.html#conjoint-experiments
}
