% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conjoint-experiments.R
\name{conjoint_measurement}
\alias{conjoint_measurement}
\title{Conjoint experiment assignment handler: conducts complete random assignment of all attribute levels}
\usage{
conjoint_measurement(data, utility_fn)
}
\arguments{
\item{data}{A data.frame}

\item{utility_fn}{a function that takes data and returns an additional column called U, which represents the utility of the choice}
}
\value{
a data.frame
}
\description{
See https://book.declaredesign.org/experimental-descriptive.html#conjoint-experiments
}
