% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-difference-in-differences.R
\name{did_multiplegt_tidy}
\alias{did_multiplegt_tidy}
\title{Tidy helper function for did_multiplegt}
\usage{
did_multiplegt_tidy(data, ...)
}
\arguments{
\item{data}{a data.frame}

\item{...}{options passed to did_multiplegt}
}
\value{
a data.frame of estimates
}
\description{
Runs did_multiplegt estimation function and returns tidy data frame output
}
\details{
See https://book.declaredesign.org/observational-causal.html#difference-in-differences
}
