% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch17_helpers.R
\name{estimator_AS_tidy}
\alias{estimator_AS_tidy}
\title{Tidy helper function for estimator_AS function}
\usage{
estimator_AS_tidy(data, permutatation_matrix, adj_matrix)
}
\arguments{
\item{data}{a data.frame}

\item{permutatation_matrix}{a permuatation matrix of random assignments}

\item{adj_matrix}{an adjacency matrix}
}
\description{
Runs estimates estimation function from interference package and returns tidy data frame output
}
\details{
See https://book.declaredesign.org/experimental-causal.html#experiments-over-networks
}
