% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-experiments-over-networks.R
\name{estimator_AS_tidy}
\alias{estimator_AS_tidy}
\title{Tidy helper function for estimator_AS function}
\usage{
estimator_AS_tidy(data, permutatation_matrix, adj_matrix)
}
\arguments{
\item{data}{a data.frame}

\item{permutatation_matrix}{a permutation matrix of random assignments}

\item{adj_matrix}{an adjacency matrix}
}
\value{
a data.frame of estimates
}
\description{
Runs estimates estimation function from interference package and returns tidy data frame output
}
\details{
The estimator_AS_tidy function requires the 'interference' package, which is not yet available on CRAN.

To use this function:
\enumerate{
\item install the developer version of interference via remotes::install_github('szonszein/interference') and
\item install the developer version of rdss via remotes::install_github('DeclareDesign/rdss@remotes')
}

See https://book.declaredesign.org/experimental-causal.html#experiments-over-networks
}
