% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_num}
\alias{format_num}
\title{Round and pad a number to a specific decimal place}
\usage{
format_num(x, digits = 3)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of digits to retain}
}
\value{
a character vector of formatted numbers
}
\description{
Round and pad a number to a specific decimal place
}
\examples{

std.error <- c(0.12, 0.001, 1.2)
format_num(std.error)

}
