% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch17_helpers.R
\name{get_exposure_AS}
\alias{get_exposure_AS}
\title{Helper function to obtain the observed exposure for the Aronow and Samii estimator}
\usage{
get_exposure_AS(obs_exposure)
}
\arguments{
\item{obs_exposure}{A numeric vector}
}
\value{
a data.frame of observed exposure to a treatment
}
\description{
See https://book.declaredesign.org/experimental-causal.html#experiments-over-networks
}
