% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_replication_file.R
\name{get_rdss_file}
\alias{get_rdss_file}
\title{Download a replication file from the dataverse archive for Research Design in the Social Sciences: Declaration, Diagnosis, and Redesign}
\usage{
get_rdss_file(name)
}
\arguments{
\item{name}{quoted name of the file on the dataverse archive}
}
\value{
an r object
}
\description{
See https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/HYVPO5 for a full list of available files
}
\examples{

\donttest{
diagnosis_2.1 <- get_rdss_file("diagnosis_2.1")
diagnosis_2.1
}
}
