% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lapop_brazil}
\alias{lapop_brazil}
\title{Data used in student exercises for RDSS based on LAPOP survey of Brazil in 2018}
\format{
A data.frame
}
\usage{
lapop_brazil
}
\description{
These data were resampled with replacement from LAPOP data (to 10,000 rows) for a subset of variables. These data cannot be used for scientific inferences, and are only useful for teaching purposes. ID numbers were scrambled so that individuals and municipalities cannot easily be identified.
}
\details{
Download the original data from https://www.vanderbilt.edu/lapop/raw-data.php

See https://www.vanderbilt.edu/lapop/core-surveys.php for survey questionnaire
}
\keyword{datasets}
