% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-regression-discontinuity.R
\name{rdrobust_helper}
\alias{rdrobust_helper}
\title{Helper function for using rdrobust as a model in \code{declare_estimator}}
\usage{
rdrobust_helper(data, y, x, subset = NULL, ...)
}
\arguments{
\item{data}{a data.frame}

\item{y}{unquoted name of the outcome variable}

\item{x}{unquoted name of the running variable}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process}

\item{...}{Other arguments to \code{rdrobust}}
}
\value{
rdrobust model fit object
}
\description{
Helper function for using rdrobust as a model in \code{declare_estimator}
}
