% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdss.R
\docType{package}
\name{rdss}
\alias{rdss}
\title{rdss package}
\description{
Companion datasets and functions for the book "Research Design in the Social Sciences: Declaration, Diagnosis, and Redesign" (book.declaredesign.org)
}
